dev: merged retry tests to a single file
